requirements.txt
aiohttp==3.8.1
djangorestframework==3.12.4
Django==3.2.9
python-decouple==3.5